refactor: domain패키지와 view 패키지 분리
