+++ src/main/java/racing/domain/Cars.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 import java.util.*;
 import java.util.stream.Collectors;
