+++ src/main/java/racing/domain/Name.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 import java.util.Objects;
 
 public class Name {
