+++ src/main/java/racing/domain/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 public interface RandomNumberGenerator {
     int getRandomNum();
