+++ src/test/java/racing/GameRunnerTest.java
@@ -2,6 +2,9 @@ package racing;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.domain.Car;
+import racing.domain.Cars;
+import racing.domain.GameRunner;
 
 import java.util.ArrayList;
 import java.util.List;
