+++ src/test/java/racing/RandomNumberGeneratorTest.java
@@ -3,6 +3,8 @@ package racing;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
+import racing.domain.CarRandomNumberGenerator;
+import racing.domain.RandomNumberGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
