+++ src/test/java/racing/TestRandomNumberGenerator.java
@@ -1,5 +1,7 @@
 package racing;
 
+import racing.domain.RandomNumberGenerator;
+
 public class TestRandomNumberGenerator implements RandomNumberGenerator {
     private int number;
     public TestRandomNumberGenerator(int number) {
